% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vasarely.R
\name{plot_vasarely}
\alias{plot_vasarely}
\title{plot_vasarely}
\usage{
plot_vasarely(vasarely_chart)
}
\arguments{
\item{vasarely_chart}{vasarely chart which has been calculated by the function vasarely}
}
\value{
returns the plotted vasarely chart
}
\description{
plot_vasarely produces the plot of a vasarely chart and allows to add further functions of ggplot2.
}
\examples{
# create data
a1 <- c(rep("A", each = 25), rep("B", each = 75))
a2 <- c(rep("A", each = 50), rep("B", each = 50))
data <- data.frame(a1, a2)

# use function vasarely to calculate list with chart
vasarely_plot <- vasarely(data = data)

# put plot into plotting function and add further function of ggplot2
plot_vasarely(vasarely_plot) + labs(title = "My vasarely chart")
}
